Make Proposition more structured
