+++ src/nlug.ts
@@ -1,4 +1,5 @@
 import { Move } from "./types";
+import { objectsEqual} from "./utils";
 
 interface NLUMapping {
   [index: string]: Move;
@@ -12,7 +13,7 @@ const nluMapping: NLUMapping = {
   },
   "What's your favorite food?": {
     type: "ask",
-    content: (x) => `favorite_food ${x}`,
+    content: (x) => ({"predicate": "favorite_food", "argument": x}),
   },
   Pizza: {
     type: "answer",
@@ -40,41 +41,16 @@ const nlgMapping: NLGMapping = [
   [
     {
       type: "answer",
-      content: `favorite_food pizza`,
+      content: {"predicate": "favorite_food", "argument": "pizza"},
     },
     "Pizza.",
   ],
 ];
 
-function deepEqual(obj1, obj2) {
-    if (obj1 === obj2) {
-        return true; // same reference or both are null/undefined
-    }
-
-    if (typeof obj1 !== 'object' || typeof obj2 !== 'object' || obj1 === null || obj2 === null) {
-        return false; // primitive values or one of them is null
-    }
-
-    const keys1 = Object.keys(obj1);
-    const keys2 = Object.keys(obj2);
-
-    if (keys1.length !== keys2.length) {
-        return false; // different number of properties
-    }
-
-    for (let key of keys1) {
-        if (!keys2.includes(key) || !deepEqual(obj1[key], obj2[key])) {
-            return false; // different properties or values
-        }
-    }
-
-    return true;
-}
-
 export function nlg(move: Move | null): string {
   console.log("generating...", move);
   const mapping = nlgMapping.find(
-    (x) => deepEqual(x[0], move),
+    (x) => objectsEqual(x[0], move),
   );
   if (mapping) {
     return mapping[1];
